requirements.txt
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
pyarrow==19.0.1
jupyter==1.1.1
